Create doc module that handles downloading and rendering. Move screens here.
