+++ doc/__init__.py
@@ -1,0 +1,2 @@
+from thread import ThreadDoc
+from thread_screen import ThreadScreen
